requirements.txt
Django>=4.2
djangorestframework
Pillow
gunicorn
dj-database-url
psycopg2-binary
cloudinary
django-cloudinary-storage